Convert embedding SDK shared files to TypeScript
