+++ tsconfig.json
@@ -26,6 +26,7 @@
   },
   "include": [
     "frontend/src/**/*",
+    "frontend/build/**/*",
     "enterprise/frontend/src/**/*",
     "frontend/test/**/*",
     "enterprise/frontend/test/**/*"
